Refactoring EventDefinitionForm to functional component and TS
